created new files in the Variant Info package
