--- components/service/myvariantinfo/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "myvariantinfo",
-  "version": "1.0.0",
-  "license": "CC-BY-NC-SA-4.0",
-  "author": "Daniel J. B. Clarke <danieljbclarkemssm@gmail.com>",
-  "private": true,
-  "dependencies": {
-    "react": "^18.2.0"
-  },
-  "devDependencies": {
-    "@types/react": "^18.0.21"
-  }
-}
